Added Solutions for Day 2.
General solution could also work by creating a dict with the amounts of colored cubes in each game. Part 1 would be checking if the max of each color is not higher than the possible amount, Part 2 would be multiplying the max of every color to get the value for the game.